Move sortedlist to mips/normrange/query/util.
